:recycle: Split factory helpers out into separate files.
Why? To make room for some new ones.
