+++ api/tests/factories/helpers/constants.ts
@@ -1,0 +1,1 @@
+export const POSTGRES_INT_4_MAX = 2147483647
