+++ api/tests/factories/helpers/presence.ts
@@ -1,0 +1,5 @@
+export function presence<T>(value: T | undefined, defaultValue: T): T {
+  return value === undefined ? defaultValue : value
+}
+
+export default presence
